Fix output-nonbranching with MinimizePN
There is an assertion that makes sure that output-nonbranching was already
handled before the SMTInterpolHelper is created. This assertion was firing,
because the original properties were passed around and thus it wasn't visible
that output-nonbranching was already handled.

This commit changes the code so that getLocationMap() modifies the PNProperties
instance directly and changes all callers to handle this correctly.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>