SMTInterpolHelper: Fix distributability
If a region should be distributable, then it should either belong to location 1
or location 2 or... When a region belongs to location i, this means that for all
events with different locations, no tokens are consumed by the event.

However, when converting this code from using an inequality system to using
SMTInterpol directly, I accidentally skipped this "or" and instead required that
the resulting region is in all locations. If there is more than one location,
this means that no event may consume any token at all, effectively making all
event/state separation problems unsolvable.

Fix this by re-introducing the missing "or".

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>