adt: Don't copy sets before returning them
This is just a useless copy. In the few cases that the caller wants to iterate
over these while making modifications, they should make a copy themselves. In
fact, only a single, artifical test in our test suite does this.

This commit introduces a new class CollectionToUnmodifiableSetAdapter. Our
datastructures have Maps whose values are unique, but since the general map
contract doesn't guarantee this, we only get a Collection instead of a Set. The
CollectionToUnmodifiableSetAdapter class turns this into a Set again.

After this commit, only the getEdges() functions still create a copy. This is a
little more complicated to deal with and will be done later.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>