Integrate findbugs into build.xml
Before this, we provided a build.findbugs.xml which one had to include from his
build.local.xml before using it.

With this commit, you just have to provide the path to the findbugs home. This
can e.g. be done by calling ant with ant -Dfindbugs.home=/path/to/findbugs.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>