ModuleParameterVerifyTask: Also follow superclasses
This splits the ModuleParameterVerifyTask into two phases. In the first phase,
information about all classes is collected. In the second phase, the classes are
analyzed for mis-use of the Module API.

Thanks to these two phases, we can now take into account information about the
superclass in the second phase. In fact, the code now merges the information of
a class with the information about all superclasses. This allows to analyze the
code of modules correctly which inherit from something that helps them do their
job.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>