ModuleParameterVerifyClassVisitor: Analyze all methods
Before this, the code only did specific analysis. For example, optional
parameters could only be declared in a method called "require".

This commit changes the code so that instead all methods are analyzed. The
assumption here is that the names of the methods that are called are unique
enough so that we do not have to check that they are only used in the
provide()/require()/run() method.

Thanks to this commit, we no longer generate warnings for the various liveness
modules which are all based on AbstractLiveModule.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>