Improve Marking memory usage and speed
This commits adds a List<Place> to the Petri net which contains all places. This
list is then references by Marking instances (not copied!) to assign an index to
each place. The actual tokens are then saved in another List<Token>.

The token list in the Petri net is never modified. Instead, if a new list is
needed, first a copy of the old one is made. This makes it possible for
Marking#ensureConsistency() to still access the old list and thus everything is
fine (tm).

This was originally done to improve the memory usage of Markings, because
VisualVM showed that the HashMap used previously caused high memory usage
(Map.Entry instances and a fill factor of 0.75...).

To measure memory usage, the CoverabilityModule was modified to add a println()
and a Thread.Sleep() after the output was fully calculated. Then, VisualVM was
used to run the garbage collector and take a heap dump. In this heap dump, the
retained size for all Marking instances was measured. Thus, the measurements
include all Markings for the reachable state space of the Petri net.

Coverability graphs were generated for (first number: memory usage before this
patch; second number after this patch):

- bitnet 14                 59 MiB   10 MiB
- quadstate_philosopher 8   65 MiB   11 MiB
- cycle_generator 40 3      53 MiB    9 MiB

However, this also has a positive effect on the run time for the coverability
graph (wall clock time, measured with /bin/time):

- bitnet 16                  20 s   11 s
- quadstate_philosopher 10   36 s   17 s
- cycle_generator 80 3       62 s   20 s

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>