Connectivity: Replace recursion by iteration
This replaces the recursion in handleStronglyConnectedComponents() by an
iteration instead. This fixes stack overflows for medium-sized LTS. The
implementation has a callers list which contains the information that was
previously saved via recursion.

Instead of also saving the postset iterator to continue where we previously
stopped, this just re-processes the postset later on.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>