PetriNet: Test creation of flows between two places / two transitions
This adds two new tests to the PetriNetTest that try to create a flow between
two places and between two transitions. Of course this shouldn't be possible,
but currently the data structures allow to do this.

The next commit will fix this, but this commit already adds the new exception
that will be thrown, so that the test can already use them.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>