Update the bundled SMTInterpol version
With the previous version, APT dies from an OutOfMemoryError in ./apt.sh
word_synthesize none :aabcccabbacbbba:. This occurs while trying to separate
event a from state s3.

For this call, the following SMT-input was given to SMTInterpol. The problem can
be reproduced via this.

  (set-logic QF_LIA)
  (declare-fun var0 () Int)
  (declare-fun var1 () Int)
  (declare-fun var2 () Int)
  (assert (and
    (> 0 (+ (* 2 var0) var2))
    (> 0 (+ var0 var2))
    (> 0 (+ (* (- 2) var0) (* (- 4) var1) (* (- 5) var2)))
    (> 0 (* (- 3) var1))
    (> 0 (+ (* (- 1) var0) (* (- 3) var1) (* (- 2) var2)))))
  (check-sat)

A bisect found out that the fix is:

  commit db5d02a36e797c76b7bee0d0146b468df622f741
  Author: Jochen Hoenicke <hoenicke@gmail.com>
  Date:   Sat Nov 29 14:03:34 2014 +0100

      Disabled branching on linear variables.

      My benchmark tells me that this has a negative effect on QF_LIA.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>