FindWordsModule: Use String instead of List<String>
This module deals with words where every event is a single character. However,
these are saved as a List<String> where each string has length one. This commit
replaces this and instead uses a String to save these words.

The reason for this is memory efficiency: Over a big alphabet, the list of
solvable words that this module computes is the only thing that stays
permanently in memory. If each word is saved as an ArrayList<String>, then the
underlying Object array uses all the memory. With Strings, we have a char array
and that should only need half the memory.

(Next steps would then be to replace the LinkedHashSet with e.g. an ArrayList
and using a binary search in there for checking if a word is solvable. However,
that would require changing the order in which we generate words...)

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>