FindWordsModule: Use an ArrayList instead of a HashSet
A HashSet has a big memory overhead, because it internally uses a HashMap which
allocates an array that should always have 25% empty entries and whose entries a
Map.Entry instances which need additional memory. An ArrayList on the other hand
is almost as good as it can be: The array is completely usable for the content
and no extra objects are allocated.

Since the list of solvable words is what needs the most memory in the
FindWordsModule and which lives for the longest time, it saves a lot of memory
to use an ArrayList here, especially for big alphabets.

This commit changes the code to use an ArrayList. Instead of using .contains()
for checking if a word is contained (which word be relatively slow), this uses a
binary search instead. Also, the code is changed so that the words are generated
in a sorted order, so that no sorting is needed. An assertion ensures that this
sorted order is not violated.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>