AbstractRenderer: Close the created writers/streams
...by using the try-with-resource statement.

Note that in renderFile() the OutputStream os might be closed multiple times,
because it gets closed directly, but also closed when the other classes chain up
and call their parent's close(). However, for Closable (not AutoClosable!), the
interface requires the close() method to be idempotent and so this is safe.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>