AllSmallCyclesHavePVOneTest: Add a failing test
This test is a lts with two states and two events that looks like this:

        a,b
      /-----> *
  -> * <-----/
         a,b

(Every event is always enabled and always goes to "the other" state.)

This test disproves an assumption that AllSmallCyclesHavePVOne does: It assumes
that its preconditions imply that all small cycles either have the same or
disjoint Parikh vectors. This is because its small cycles have PVs (1,1), (2,0)
and (0,2).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>