Reorganize AllSmallCyclesHavePVOneTest
In the current version, the code that is being tested is always run through
checkCycles(). This function asserts that if smallCyclesHavePVOne() returns
false, then no cycle with PV one has been found. However, it is possible that a
cycle with PV one is found and later also a counter example "shows up". Thus,
this assertion is wrong.

This commit splits the code up into several different methods for checking the
result. Each such method does the correct assertions directly.

For testWithUncomparableCycle(), no helper function is used at all.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>