build.xml: Fix classpath and compilation for javac.compiler
The output path of the compiler is apparently implicitly part of the class path.
So if some file was modified, the META-INF directory already existed from the
previous invocation, the <depend> task would delete the annotation processor and
then the compiler would fail because it tried loading a non-existent annotation
processor.

Fix this by removing the META-INF directory before compilation and by not having
the full project class path in here at all (where it just does not belong).

Reported-by: vsp <vsp@chuck.Informatik.Uni-Oldenburg.DE>
Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>