build.xml: Call the Java compiler via a macro
While not helping much with the size of the file, I think that this improves the
readability a lot. It is now almost trivial to spot the differences between the
various compiler invocations while previously it was hard to see what was
changed.

This macro automatically guesses the classpath to use based on the name of the
target. To make this work, test.class.path has to be renamed to
tests.class.path.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>