.travis.yml: Add a simple "smoke test"
This calls the apt.jar after it was build and just checks that it does return a
non-zero exit status when asked to "draw" some APT file. This verifies that the
module list finds (at least) the draw module and also that the APT parser and
the DOT renderer is found. So, it's just a simple smoke test to see if APT is
completely broken.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>