Optimize the ParikhVector class
This applies the same optimization that was already done to the Marking class to
the ParikhVector class. Namely, replace the Map which has memory overhead, e.g.
due to the Map.Entry instance, and which isn't really the fastest option,
because most operations work on all entries. This map is replaced with two
ArrayLists. One contains the labels and the other list contains the number of
occurrences for the label with the same index. Assuming that the number of
labels is relatively small, this should be an improvement in all cases.

The time to run compute_pvs on the reachability graph of a four tristate
philosophers net improves from ~20 sec to 7.2 sec (FloydWarshall) respectively
from ~9 sec to 3.5 sec (DFS).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>