SubGraph: Add methods for further "subgraphing"
The difference to the existing methods for creating a subgraph lies in the graph
which the subgraph will consider its parent. These new methods create "flat"
subgraphs in the sense that the new subgraph will have the original graph as its
parent, instead of having the subgraph of the original graph as its parent.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>