Connectivity#handleStronglyConnectedComponents(): Optimize
For almost every node that this algorithm visit, it checks if this node was
already visited on the current path. This was done via LinkedList#contains()
which has quite a bad performance (O(n) and a big factor).

This commit uses a HashSet for this operation instead.

With this change, checking if the reachability graph of 9 quadstate philosophers
is strongly connected now only takes 8 seconds. It took 133 seconds before this
change.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>