checkSimplyLive(): Work directly on the coverability graph
Instead of converting the reachability graph to a TransitionSystem, this does
the needed check directly on the coverability graph. This also removes the
UnboundedException that could be thrown. There is no problem here, because the
coverability graph is enough to test if a transition is simply life or dead.

This speeds up the simply_live check on a 9 quadstate philosophers nets from 30
seconds to about one third of a second.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>