Token: Add a token cache via Token#valueOf()
Instead of creating a new instance of Token whenever one is needed, it makes
more sense to re-use instances, as is done via Integer#valueOf. This especially
makes sense, because Petri nets will usually only generate a small number of
different token counts, but these will be refered to often.

This commit adds a static array to Token that is initialized with token
instances from 0 to 255. The valueOf() function uses this array if its argument
is in range and calls the normal constructor if not.

This cache is filled from a static initializer so that we don't need a runtime
check for not-yet-initialized cache entries and because the code stays
thread-safe this way. 256 instances of Token don't need much memory anyway.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>