Deprecate public Token constructor
This commit updates all code to use Token#valueOf() instead. When generating the
coverability graph of a 16 bit Petri net, this change saves about 3 MiB of
memory for the 131.088 different instances of Token that were created before.
Thanks to the cache in Token#valueOf(), all of these now share Token instances.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>