CoverabilityGraph: Remove unvisited Deque
The Deque contained all nodes for which we possibly did not yet generate the
postset. However, we already have a list of all nodes that we know that is in
the same order as this Deque. Thus, it can be replaced with a simple index into
the list of all nodes. All entries after this index are unvisited.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>