build.xml: Remove depend-classpath attribute from javac-part
Instead of specifying the depend classpath everywhere and possibly introducing
errors, we can just always specify a possibly too large classpath of classes/*
for the <depend> task. It doesn't matter since this task only figures out which
files have to be recompiled, so too few entries is a problem, but too many is
ok.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>