Rewrite the ParikhVector class
The biggest difference is that this commit gets rid of a ParikhVector's
connection to a TransitionSystem. Trying to do anything sensible when the
underlying TS is modified is hard, because most likely the ParikhVector loses
the meaning that it previously had. Thus, ParikhVectors now exist on their own
without any underlying TS.

Also, instead of lessThan(), this provides tryCompareTo() which works similarly
to Comparable#compareTo(). However, uncomparable ParikhVectors also produce a
result of 0. To check for this, the API user can use the new isUncomparableTo()
method that this commit adds.

Because the ParikhVector is no longer connected to a TransitionSystem, it also
has no concept of an alphabet. Thus, entries with a count of 0 are discarded,
which causes a different result in toString(). Unit tests are fixed up to deal
with that, but honestly, those unit tests shouldn't rely on toString().

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>