ComputeSmallestCycles: Fix bug captured in the new unit test
The trick is to remove *all* larger cycles when a smaller one is found, not only
the first one that we come across. To make sure this kind of problem doesn't
happen again in the future, there is now only one implementation of this that is
used by all three algorithms.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>