Add a service-verify task
After building a JAR file, this task is run. It will load several classes from
the JAR file that was just produced and check if any exceptions occur. These
catches cases where e.g. a module is removed, but not removed from the module
list (we have no automatic step for this).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>