PetrifyPNParser: Add support for more than .inputs
Petrify comes from a circuit design context. Here, .inputs are signals that
aren't controlled by the circuit, but come from the outside. .outputs are
signals to the outside. .internal are internal signals and .dummy describes
silent signals.

These differences do not much to us and we will treat all of these the same.

Thanks to this, APT can now read petrify files generated by genet (which
generates an extension of the Petrify file format).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>