PetrifyPNParser: Add support for initial token
In the Petrify file format, the initial marking of each place is either zero or
one. Genet extends this to non-safe nets and here the initial marking is given
with an equal sign and a number, e.g. "p0=3". This commit adds support for this
extension of the file format.

It was tested with the following command which uses genet to synthesize a Petri
net in which transition a can fire twice. The resulting net has a single place
from which a consumes tokens. The place's initial marking is 2.

    ./apt.sh reg aa - | ./apt.sh lts_convert apt petrify - - > /tmp/t.aut &&
    genet /tmp/t.aut -min 3 | ./apt.sh pn_convert petrify apt - -

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>