Base the PetrifyPNRenderer on the GenetPNRenderer
The old renderer was buggy in various ways that I won't mention here. However,
since the Genet file format is an extension of the Petrify file format, we can
just re-use the GenetPNRenderer and make sure that the resulting file conforms
to the Petrify file format (only arc weight 1, only single tokens in the initial
marking).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>