Add a GenetPNParser
Our PetrifyPNParser was extended so that it can also read genet files. However,
our renderer should complain if a Petri net cannot be expressed in the petrify
file format and thus a new implementation was written. This means that the list
of parsers claims that we can only generate the Genet file format, but not parse
it.

This commit adds a GenetPNParser that just uses the PetrifyPNParser to overcome
this.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>