Change the optional parameters in ModuleInputSpec
Right now, only the default value can be set for an optional parameter. The
module help output uses the toString()-representation of this value. This works
for the types that use this right now (String, Character, Integer), but isn't
really expandable to any "complex" types.

Thus, this commit splits addOptionalParameter() up into two variants, one
without a default value and another with a default value. This API change makes
it easier to find spots that need to be changed (since with a String.class
argument something could silently break) and makes it clearer that an argument
doesn't have a default value (null is quite a hack for this).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>