AbstractMapParameterTransformation: Allow using unique prefixes
This commit changes the AbstractMapParameterTransformation so that it checks if
the given argument is either contained in its map (it already did this before)
or if it's a unique prefix of an entry (this is new). This keeps things like
"next" vs "next_valid" working, while making it possible to have more meaningful
arguments for things like selecting "farkas" algorithm instead of algorithm "f".

Thanks to vsp for the suggestion.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>