Fix 'ant xml.test' on Windows
Just because we want to group tests by package, we have lots of magic that
generates an input file for TestNG describing all tests in such a grouped way.
This magic gets a file name and needs to figure out the package and class.

For this, it uses String.split(File.separator). On Unix, where File.separator is
"/", this works fine. However, on Windows the single backslash is not a valid
backslash and this code fails to get the components of the file path.

To fix this, this code now uses java.io.File to parse the file path and to
iterate over it. This should work on all platforms.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>