Add a module for "Petri net resynthesis"
This module has arguments similar to the already existing synthesis module
(SynthesizeModule). However, it also gets an additional Petri net. The places of
this Petri net are transformed into regions and these regions are given to
SynthesizePN. The effect is that, if PN synthesis is successful, the resulting
Petri net will re-use as many places from the input as is sensible.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>