FindWords: Use threads for finding solvable words
This makes the code use a thread pool with as many threads as the system has
processors. Each thread solves a word by creating the SynthesizePN instance for
some word and then returns this instance to the main thread. The main thread
then handles the solvable/unsolvable result. This way, users of this code do not
have to concern themselves with thread-safety.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>