FindWords: Avoid OutOfMemoryError
Before this commit, the code created all Callables for a given word length up
front, submitted them for execution and then waited for all of them to be done.
Since the number of words grows exponentially (does it?), this causes
OutOfMemoryErrors when longer words are reached.

To avoid this, create Callables lazily. The Iterator that the previous commit
added is used to generate words that need to be checked for solvability. A
TransformedIterator is then used to get an iterator that returns Callables for
execution. This iterator is then used to submit new jobs for execution whenever
the number of pending jobs is low, that is when the number of jobs which are
queued because all worker threads are currently busy is low.

The only thread pool that I found that allows to query the number of queued
submissions is the ForkJoinPool. Thus, the code now uses this class. A nice
result of this is that ForkJoinPool provides a no-argument constructor which
creates a pool with as many worker threads as the system has processors and so
we no longer have to query this value ourselves.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>