Rewrite the AptLTSRenderer with StringTemplate
This adds support for extensions to the renderer. However, it also changes the
generated output, because the initial state is now marked with [initial="true"]
instead of just [initial]. The parser already supported this before (it ignores
the value of the "initial" property), so this doesn't count as a change in the
output format.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>