FindWords: Reject the combination of plain + k-marking
For non-k-bounded Petri nets, every prefix and suffix of a solvable word can be
solved. For suffix, just fire the first transition to get a new initial marking.
For prefix, add for each letter a counting place from which the corresponding
transition consumes one token. A minimally unsolvable word is thus a word which
is itself unsolvable and where all proper subwords are solvable.

For k-bounded Petri nets, every suffix is solvable in the same way while
prefixes might be unsolvable. A minimally unsolvable word is thus an unsolvable
word where each proper suffix is solvable.

However, this fails when plain + j-marking is required. With just j-marking we
can fix things by multiplying the initial marking and all arc weights with the
factor j. This does not work when additionally plain is required. This can in
fact not be worked around, because with j>1 the word "a" is unsolvable while a^j
(j-times the letter a) is solvable. In this setting the whole approach of the
FindWords code breaks.

This commit makes FindWords reject this situation.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>