AptPNRenderer: Support extensions
The renderer first checks if any flow has an extension. If not, then the same
format as before is generated. Otherwise, each flow is printed in its own line
together with its extensions. This means that a transition is mentioned multiple
times in the file, but the parser has no problems with that and always supported
it.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>