Reintroduce old Extensible#putExtension() method
There are two reasons for this and both are bad reasons:

- This new method is a tiny bit faster, because no empty array is created
- This fixes compatibility with old versions

To explain the second point a bit more: Apparently ant's depend task does not
detect the switch from putExtension(String key, Object value) to
putExtension(String key, Object value, ExtensionProperty... properties)
correctly. Thus, callers of this method are not recompiled and instead a
NoSuchMethodError occurs at run-time.

Users can fix this by running "ant clean" to force a recompile, but this is bad
since it requires us to tell everyone about it when he runs into this problem.
This commit instead makes sure that the old bytecode continues to work.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>