Fix ParikhVector#tryRemove()
This method is documented to return null if the ParikhVector has fewer entries
than the number that is requested to be removed. However, the actual
implementation succeeded as long as the ParikhVector contained the given event
at least once, because Bag#remove() returns true whenever the Bag changed as a
result to this call (= when the Bag contained the entry at least once).

This wasn't noticed previously, because the only caller of this function has
hardcoded the argument n=1. In this case the old implementation worked
correctly.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>