Speed up TransitionSystem's event management
The transition system has an alphabet which contains all events which are used
to label some arcs. The current implementation for this is quite inefficient,
because getEventInternal() does a linear search through the known events to find
one with a given label. Due to this, calculating the isolated elements of the
reachability graph of a 16 bit net takes more than 10 minutes (and most of this
time is spent for reading the input file and creating a TransitionSystem
instance from it):

$ /bin/time ./apt.sh isolated_el /tmp/bit-16-reach.apt
isolated_elements: []
677.72user 0.94system 10:49.07elapsed 104%CPU (0avgtext+0avgdata 3003888maxresident)k

Speed this up by using a Map from the string used for labels to the Event
instances to manage the alphabet. Since we also need the alphabet as a set, an
additional set is used. To save the number of references, an InternalEvent class
is added which manages a reference counter. Adding an arc with label "e"
increases this reference counter and removing the arc decreases it again.

With these changes, the reachability graph of a 16 bit net can be read in less
than 10 seconds:

$ /bin/time ./apt.sh isolated_el /tmp/bit-16-reach.apt
isolated_elements: []
35.39user 0.81system 0:09.46elapsed 382%CPU (0avgtext+0avgdata 2728692maxresident)k

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>