Fix the build on Java 7
TransitionSystem used to guarantee that its alphabet is sorted according to the
natural ordering of the events. The previous commit broke this by using a
HashSet which does not provide this natural ordering. This caused some test
failures for tests that assumed this ordering.

I have no idea why this only failed on Java 7, but ok...

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>