AllSmallCyclesHavePVOne: Use better data structures
Apparently HashBag#containsAll() is implemented in quite a slow way. This commit
switches the code from a HashBag containing the events that were already fired
to a HashSet containing events which were not yet fired on the current path.

This speeds up Petri net synthesis using MarkedGraphSeparation of the
reachability graph of a connected 7 bit net from 128 seconds to 23 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>