Speed up AllSmallCyclesHavePVOne
This idea is based on work by Eike Best and Raymond Devillers who came up with
an algorithm that computes the Parikh vectors of all small cycles. This
implementation now uses a specialised version of their algorithm for checking
the P1 property.

Before this change, synthesising the reachability graph of a connected 7 bit net
with the MarkedGraphSeparation implementation took 23 seconds. This commit
speeds this up to 0.31 seconds, which shows that all of the time was spent
checking for the P1 property via AllSmallCyclesHavePVOne.

A downside of this new algorithm is that it can no longer return a cycle with a
Parikh vector incomparable to the all-ones-vector. However, except for the
tests, nothing used this anyway.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>