Petrify parser: Add support for the .capacity directive
When GENET produces a Petri net which is not 1-bounded, it (apparently) adds
information about the bound on the number of tokens of each place (called
'capacity'). Our parser does not understand this information and bails out.

This commit adds support for this information to our parser (by silently
ignoring it).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>