Modified AbstractConvertModule to work with the new parameter changes.
With the new changes the parameter type is used to decide how command
line arguments are interpreted (as raw value or file path). Strings are
always interpreted as raw values. However the AbstractConvertModule
needs the contents of a file as its input in the form of a string. To
get around this limitation unfortunately a new type had to be introduced
that is just a wrapper around String. However it is interpreted as a
file path on the command line.