Add a failing synthesize test with locations
For solving a state separation problem, we need a region for which the two given
states have a different number of tokens. A lot of the code assumes that it is
ok to strengthen this to "for which the first region has less token than the
second region". The argument for this is that for any region there is a
corresponding complement region.

However, with locations (or with property "output-nonbranching") this is no
longer true, as the test that this commit adds shows.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>