BasicPureSeparation: Fix the new, failing test
If a state separation problem seems to be unsolvable and we have locations, then
we now also attempt to synthesize a region solving the "complement" problem.
This is because we strengthen "!=" to ">" for the values that the region should
reach, but this may make a solvable problem unsolvable when locations are
involved.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>