SpanningTree: Make sure our cache is not copied
Normally, the TransitionSystem's copy constructor copies all extensions.
However, the cache that is internally used by SpanningTree should not be copied.
Fix this by adding a NOCOPY property to this extension.

The test has to modify the transition system, because the cache here is a Map
from States to a spanning tree beginning with that state. This map does by
itself not cause any problems. However, the StructuralExtensionRemover is not
added to the copy of the lts and thus the cache isn't properly cleared when
needed. This is what the test tests for.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>