OverapproximatePN: Restrict the supported classes of nets
Most properties cause an endless loop if they are used to over-approximate the
LTS (0,a,1) (0,b,1). The exception is "homogeneous", but "homogeneous,pure"
again causes an endless loop. In fact, all properties have this problem which
aren't "symmetric" in the sense that if a region satisfies the property, then
also the complement of the region satisfies this property. For symmetric
properties, I have an argument for the correctness of the algorithm.

The implementation now only accepts properties which are symmetric in the above
sense (and for all other properties the above also provides a counter-example
for the correctness of the algorithm).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>