OverapproximatePN: Fix a bug with k-bounded synthesis
It is possible that a k-bounded region is no longer k-bounded after a new state
is added to a transition system. For example, imagine the lts that has just an
initial state and the region with initially no tokens and transition 'a' adds 10
tokens. If now a new state is added which is reachable from the initial state
via 'a', the region which was previously 0-bounded is now only 10-bounded.

This means that OverapproximatePN cannot just copy regions from one iteration to
the next iteration if k-bounded synthesis is done. Instead, it has to check if
the regions from the last iteration are also k-bounded in the new iteration (but
only if any new states were added to the lts).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>