Fix a case where unsolved SSP instances were wrong
The function calculateUnseparatedStates() calculates all states which are
involved in an unsolved SSP instance for a given set of regions. However, it
mishandled unreachable states. In the presence of an unreachable state, all
states are involved in an unsolvable SSP instance, because an unreachable state
cannot be separated from any other state. Instead of this, the function might
only have returned the actual unreachable states.

Fix this in the obvious way.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>