APT#main: Split output handling into a helper function
The code that prints the result of a module to standard output or files is long.
This commit moves it from main into a helper function. This halves the length of
main which after this commit has 90 lines while the helper function has 83
lines.

This split makes it easier to follow the code, because, for example, there are
less local variables in scope that can be used by the code and have to be kept
in mind by the programmer.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>