Fix KBoundedSeparation
By accident I noticed the following situation: If I synthesise some LTS with APT
and properties "none", I get a 5-bounded PN. If I synthesise this LTS with
properties "5-bounded", APT fails to solve the input. Weird.

It turns out that the algorithm behind KBoundedSeparation doesn't actually
guarantee success. In fact, it is not true that minimal regions are sufficient
for synthesis (this is true for elementary Petri nets).

This commit fixes KBoundedSeparation by removing the minimisation step.

This change unveils a bug in a test: OverapproximatePNTest contained a test with
3-bounded separation that was actually not testing against the minimal
over-approximation of the input. The minimal 3-bounded overapproximation of the
unsolvable word "abbaa" just has a single "a side arm" added at the separation
failure.

Since the original example that caused this problem is a lot more complicated
than "abbaa", I'm not giving more details here.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>