Implement some list-of-words -> LTS transformation
The basic idea is to define an LTS by a list of words that should be in the LTS'
prefix language.

Additionally, it is possible to provide a Transformer that can additionally
identify states with each other (i.e. the words "a","b" and "b" should reach the
same states). To simplify this, states are identified by the ParikhVector of a
word reaching them, so by itself the code will already make "a","b" and "b","a"
reach the same state.

This makes perfect sense for the intended use for Petri net synthesis.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>