Add a matcher for rotated Iterables
For example, assertThat(Arrays.asList(1,2,3), containsRotated(2,3,1))
succeeds, because the list is rotated by one (the "1" is moved from the
beginning to the end), while containsRotated(2,1,3) would fail, because
the given list of not a rotation of the input.

To make the idea of what is meant with 'rotation': The rotations of
"1,2,3" are "1,2,3", "2,3,1", and "3,1,2". The matcher matches if any of
these rotations is the same as its input.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>